SQLite3 database
Data downloaded from the Open Tree of Life API is stored as JSON so we can decide later which parts of that data we want to use. Because it takes a long time to get all these records, we want to start ASAP and not be blocked by concerns over how that data should be formatted.

I switched from CSV to a relational database for better write performance and also so we can analyze the data as it trickles in.